Extract logic into separate modules
